feat(serialization): move polymorphic serialization module to external object
