+++ src/main/kotlin/org/careerseekers/userservice/config/CacheConfig.kt
@@ -1,7 +1,7 @@
 package org.careerseekers.userservice.config
 
 import org.careerseekers.userservice.dto.CachesDto
-import org.careerseekers.userservice.dto.json
+import org.careerseekers.userservice.serializers.CustomSerializerModule.json
 import org.careerseekers.userservice.serializers.PolymorphicRedisSerializer
 import org.springframework.cache.CacheManager
 import org.springframework.context.annotation.Bean
