+++ src/main/kotlin/org/careerseekers/userservice/enums/MailEventTypes.kt
@@ -1,0 +1,6 @@
+package org.careerseekers.userservice.enums
+
+enum class MailEventTypes {
+    REGISTRATION,
+    PASSWORD_RESET
+}
